+++ crates/module-system/sov-modules-rollup-blueprint/src/endpoints.rs
@@ -5,19 +5,20 @@ use sov_modules_api::hooks::ApplyBatchHooks;
 use sov_modules_api::{RuntimeEventProcessor, Spec};
 use sov_modules_stf_blueprint::{Runtime as RuntimeTrait, RuntimeEndpoints, TxReceiptContents};
 use sov_rollup_interface::zk::{ZkvmGuest, ZkvmHost};
-use sov_sequencer::{FairBatchBuilder, FairBatchBuilderConfig, SequencerDb, TxStatusManager};
-use sov_stf_runner::SequencerConfig;
+use sov_sequencer::batch_builders::standard::StdBatchBuilder;
+use sov_sequencer::batch_builders::BatchBuilder;
+use sov_sequencer::{BatchBuilderConfig, SequencerConfig, SequencerDb};
 use tokio::sync::watch;
 
 use crate::{FullNodeBlueprint, SequencerBlueprint};
 
 /// Register rollup's default RPC methods and Axum router.
-pub fn register_endpoints<B, M>(
+pub async fn register_endpoints<B, M>(
     storage: watch::Receiver<<B::Spec as Spec>::Storage>,
     ledger_db: &LedgerDb,
     sequencer_db: &SequencerDb,
     da_service: &B::DaService,
-    sequencer_config: &SequencerConfig<FairBatchBuilderConfig<B::DaSpec>>,
+    sequencer_config: &SequencerConfig<B::DaSpec>,
 ) -> anyhow::Result<RuntimeEndpoints>
 where
     B: FullNodeBlueprint<M> + 'static,
@@ -27,6 +28,35 @@ where
     <B::OuterZkvmHost as ZkvmHost>::Guest: ZkvmGuest<Verifier = <B::Spec as Spec>::OuterZkvm>,
 {
     let mut endpoints = B::Runtime::endpoints(storage.clone());
+    let da_address = sequencer_config.da_address.clone();
+    let sequencer_router = match &sequencer_config.batch_builder {
+        BatchBuilderConfig::Standard(bb_config) => {
+            let batch_builder =
+                StdBatchBuilder::<(B::Spec, B::DaSpec, B::Runtime), B::Kernel>::create(
+                    storage,
+                    da_address,
+                    sequencer_db.read_all()?,
+                    bb_config,
+                )
+                .await?;
+            let tx_status_manager = batch_builder.tx_status_manager();
+            let sequencer = SequencerBlueprint::<B, M>::new(
+                batch_builder,
+                da_service.clone(),
+                tx_status_manager,
+                sequencer_db.clone(),
+                ledger_db.clone(),
+            );
+
+            sequencer.rest_api_server("/sequencer")
+        }
+        BatchBuilderConfig::Preferred => {
+            todo!("Preferred sequencer is not yet supported")
+        }
+    };
+
+    // Sequencer endpoints.
+    endpoints.axum_router = endpoints.axum_router.nest("/sequencer", sequencer_router);
 
     // Ledger endpoint.
     {
@@ -43,30 +73,6 @@ where
         endpoints.axum_router = endpoints
             .axum_router
             .nest("/ledger", ledger_axum_router.with_state(ledger_db.clone()));
-    }
-
-    // Sequencer endpoints.
-    {
-        let tx_status_manager = TxStatusManager::default();
-        let batch_builder = FairBatchBuilder::<B::Spec, B::DaSpec, B::Runtime, B::Kernel>::new(
-            B::Runtime::default(),
-            B::Kernel::default(),
-            tx_status_manager.clone(),
-            storage,
-            sequencer_db.clone(),
-            sequencer_config.batch_builder.clone(),
-        )?;
-
-        let sequencer = SequencerBlueprint::<B, M>::new(
-            batch_builder,
-            da_service.clone(),
-            tx_status_manager,
-            ledger_db.clone(),
-        );
-
-        endpoints.axum_router = endpoints
-            .axum_router
-            .nest("/sequencer", sequencer.rest_api_server("/sequencer"));
     }
 
     Ok(endpoints)